# TheWebApiServer/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TheWebApiServer, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "TheWebApiServer, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/problem+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/HttpValidationProblemDetails"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "TheWebApiServer, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "useCookies",
            "type": "boolean"
          },
          {
            "in": "query",
            "name": "useSessionCookies",
            "type": "boolean"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AccessTokenResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AccessTokenResponse": {
      "type": "object",
      "properties": {
        "tokenType": {
          "type": "string",
          "readOnly": true
        },
        "accessToken": {
          "type": "string"
        },
        "expiresIn": {
          "format": "int64",
          "type": "integer"
        },
        "refreshToken": {
          "type": "string"
        }
      }
    },
...
    "HttpValidationProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        },
        "errors": {
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "additionalProperties": { }
    },
...
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "twoFactorCode": {
          "type": "string"
        },
        "twoFactorRecoveryCode": {
          "type": "string"
        }
      }
    },
...
    "RegisterRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}